requirements.txt
https://github.com/KurimuzonAkuma/pyrogram/archive/dev.zip
TgCrypto
Telegraph
html_telegraph_poster
Pillow